+++ custom_components/cubecoders/__init__.py
@@ -14,7 +14,7 @@ from .coordinator import AmpDataUpdateCoordinator
 from .data import AmpData
 from .entry import AMPConfigEntry
 
-_PLATFORMS: list[Platform] = [Platform.SENSOR]
+_PLATFORMS: list[Platform] = [Platform.SENSOR, Platform.SWITCH, Platform.BUTTON]
 LOGGER: Logger = getLogger(__package__)